[UNSAFE] KVM: arm64: Add pkvm proxy module
this is a port for android15-6.6 of the original commit for pKVM on linux 6.4

This allows when enabled, to completely bypass the regular KVM API
and interact directly the hypercall API. This is usefull to test the
pKVM hypervisor. However this completely break the kernel security
model, any kernel having this commit is severly compromised.

The ioctl-based UAPI is described in
arch/arm64/include/uapi/asm/pkvm_proxy.h

The set of supported hypervisor operations is enough to boot a
non-protected guest on pkvm-core-6.4. Other version might need
modifications to reach that stage

Change-Id: I8b9aa25bbb5fabbba3fae10641a828be8511965f