Add a -i mode to the carver-wrapper script
When called with -i, it drops the user into an interactive
shell (oops, currently fixed to bash) in the carver directory,
with a single target file, and cleans up the directory when the
shell exits.

The idea is that this is a halfway useable way to call CN
with different settings (e.g. read the state file, use the debug
modes).