Mark some functions as trusted to cover more code
The bogus trusted annotations shelve problematic functions to permit
more coverage and get more of a sense of what needs fixing in
CN/cerberus. The next thing to do is come up with a better way to
handle shift-left ops, which for the moment create unpleasant
obligations about nonlinear arithmetic.