Add level-validity constraints to PT invariant
The "level" (or depth) of the current page-table is used in the
interpretation of entries. In particular, the kvm_pte_table
function always reports that entries are not tables if the max
depth is reached. For consistency, this means the invariant should
guarantee that table entries are not found within max-depth tables,
which requires tracking level information at various places.