renaming `.present` to `.exists` in `ghost_vcpu` and adding a similar field in `ghost_vm`.
In the ghost state, we currently model the table of VMs as an array with size
equal to the max number of supported VMs (instead of something resizable).
The `.exists` boolean indicates whether the VM is actually part of the logical table.

Similarly with the table of VCPUs for each VM.