(ghost/vms) remove vm_handle_offset
Now, the vms table is an unordered set, and the ghost state does not
know about the relationship between handles and indexes.

The loaded vcpu contains an opaque handle, rather than an index,
and to find the corresponding VM the vms table must be iterated over
to find the vm slot that exists with the requested handle.

Equality between vms tables is also now more complex, requiring that
the set difference between extant vms is empty.

This also means that the concept of a vm slot being "present" but
not "existing" no longer exists, and the vms table should be thought of
as a list/set of VMs. So the state either contains all the vms or none.