(ghost/ghost_spec.h) add new ghost_relaxed_reads structure
Now, there is some data to keep track of each of the relaxed reads
performed during a hypercall in addition to the already recorded
returned value.

Currently this is not used by the implementation,
need to replace the various READ_ONCE(...) invocations with a
READ_ONCE_GHOST_RECORD(...) to actually record the values.

Also, not used in the existing specifications, yet.