(ghost/ghost_spec/vms) add abstractions for managing vm table
Instead of directly indexing into the vms->tables, go via some helper
functions that manage it in a neat way: returning references to slots,
and with alloc/copy/free functions for safe management of slots.

Also keep a reference to the underlying VM lock to assert it's held
at the key points.