(ghost/hyp_va_of_host_va) add convert for host kernel VAs
We only go one way (host va -> hyp va) as we do not know the offsets for the host,
but we should never need to synthesise a host VA in the spec so that's ok.

See the comment in va_layout.c to explain the conversion.

To make this work we had to add tag_lsb and tag_val to the ghost state,
this involved making the non-nvhe-directory uses of them link to a nvhe symbol,
which is ugly but we can fix it later...

(Ben made me do this, I disclaim any liability for any future brain melting)