(ghost/compute_abstraction) pointerification
Now, each compute_* function takes a `dest` to write it into,
rather than computing large ghost structs on the stack.

Probably -O2 etc would have been smart enough to do a lot of this itself,
but we want to run with -O1 and (eventually) -O0 so we can't afford the
overzealous stack usage.