(ghost/ghost_pgtable) add pfn set (WIP, does not compile)
We extend the abstract_pgtable type with two new parts:
- the original pool range
- and the set of allocated tables outside that range

Conceptually, the locations within the pool are all
owned exclusively by pKVM, and the hypercalls enforce that,
and we track the allocations outside that pool with a set,
allowing us to specify the non-determinism in the memcache pages.

This will be done in two parts, this is part one, and does not yet compile.