(ghost/compute_abstraction) fix synchronisation
On hypercalls with multiple lock/unlock sequences,
only take snapshots of the first lock and last unlock.

This isn't correct! But, if we assume non-interference from other
hypercalls we can write a basic spec to do some basic checking.

In future, we want to replace the pre/post whole-pgtable checks
with some more granular this-thread diff calculations and checks.