(ghost/vms) remove vm_table lock checks
It is not required to hold the vm_table lock when operating on a VM,
when not indirecting via the table.

This means we can't require the vm_table lock be held during VM calls,
and this leaves the underlying ghost vms dictionary unprotected.

Probably want to add back in a ghost vm dictionary lock? Or, maybe it's
okay as it's a per-cpu struct.