(ghost/maplets) refactor
Does two major refactorings in one:
1. Enable distinguishing between stage1 and stage2 maplets.
2. Cleanly separate architectural bits with abstraction.

For 1, all maplets now carry around a stage, and the printers are stage
aware. Mostly this is a rewrite of the maplet printer and renaming of
virt and phys to ia and oa.

For 2, the changes are more drastic. Now maplets are almost entirely
architecturally agnostic, with brand new ghost data structures based
on the abstraction.  ghost_pgtable now contains parsers from concrete
architectural state to abstract maplets, but there's no way to go back,
and the maplets don't store raw protections (except for pretty-printing)

This happily also incidentally fixes a few bugs to do with Stage1 vs
Stage2 parsing of access permissions.