(ghost/printer) add ghost_printf
Overhaul of the ghost print machinery.

This adds a printf-like mechanism to print to UART or string buffers.

Generally can be used like normal printf:
> ghost_printf("%d %s:%lx\n", 42, "hello", 0x1234);

But comes with some extra printf-codes (see the comment in the header)
for ghost: %g(maplet) and %g(pgtable) for printing maplets and mappings,
and %$ for printing the shadow stack.