Rearrange mm_ops treatment, finish verification
The problem, of course, was partial coverage of two different views of the
mm_ops struct. One pointer to it, in the struct kvm_pgtable, is not visible in
every scope (this struct is not passed to the callbacks). Another version may
or may not exist in the walker data associated with the callback, as some
walkers need it and others do not. The best solution I can think of is to
expose the walker's mm_ops option and assert that it is equal to the other if
it is present. This seems to work.