(ghost/spec) improving the ghost state of vCPUs with respect their ownership
When a vCPU is initialised but not loaded anywhere, it is owned by
the `vm_table`. However when a vCPU is loaded, ownership is transferred to the
(thread-local) state of the physical CPU it is loaded on.

To model this, the `.vm_table_locked` member of a VM now holds an array
of "vCPU reference" structs which hold the metadata of the vCPU and, when the table
has ownership, the pointer to the actual vCPU object.

Similarly, the `.loaded_hyp_vcpu` member of the ghost local state now
holds a pointer to the currently loaded vCPU (or NULL when there is none).

The specs of `__pkvm_vcpu_load` and `__pkvm_vcpu_put` now move the vCPU object
back and forth between the VM state and the local state.

NOTE: still TODO ==> `.vcpu_addrs` in `vm_teardown_data` needs to be turned into
an option type, because it is owned by the vm_table.