(ghost) moving `vcpu_addr` to the `vm_table_locked` structure
This array needs to be in an option type because it is only present
when the VM table is itself present (it depends on the
number of initialised vCPUS, and we can't record it without
the VM table lock).

I tried doing that while keeping it in `struct ghost_vm_teardown_data`
but it was too much of a headache to keep the two option types in sync.