(ghost/recording) fix: local state loaded_vcpu always present
Instead of flipping pointers around between the vms table and the local
state, we have an always-present vCPU in the local state,
and simply override it.

This avoids the transient state where the pointer is NULL, but the
vCPU is known to be loaded.

Now the specs can just override the local loaded vCPU directly.